+++ HISTORY.md
@@ -19,3 +19,4 @@
 ## Version 0.2.1.a1
 * Concept.ratio added
 * Concept.size added
+* Concept.shape added